requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
langchain-groq
sentence-transformers